LayoutTests/imported/w3c/web-platform-tests/cookies/samesite/setcookie-lax.https-expected.txt


PASS Same-site window should be able to set `SameSite=Lax` or `SameSite=Strict` cookies.
FAIL Cross-site window shouldn't be able to set `SameSite=Lax` or `SameSite=Strict` cookies. assert_equals: `samesite_strict=0.8996410966959045` in `document.cookie` expected false but got true
